[refactor] 파일 폴더 구조 변경
폴더 구조를 좀 더 명확하게 하기위해 auth/security 폴더를 추가하여 그곳에 LoginUser, Filter등을 한곳에 몰아넣었음

#228
