+++ backend/src/main/java/com/ddbb/dingdong/domain/auth/service/AuthManagement.java
@@ -10,8 +10,8 @@ import com.ddbb.dingdong.domain.user.entity.User;
 import com.ddbb.dingdong.domain.user.entity.vo.Role;
 import com.ddbb.dingdong.domain.user.repository.SchoolRepository;
 import com.ddbb.dingdong.domain.user.repository.UserRepository;
-import com.ddbb.dingdong.infrastructure.auth.AuthUser;
-import com.ddbb.dingdong.infrastructure.auth.AuthenticationManager;
+import com.ddbb.dingdong.infrastructure.auth.security.AuthUser;
+import com.ddbb.dingdong.infrastructure.auth.security.AuthenticationManager;
 import com.ddbb.dingdong.infrastructure.auth.encrypt.PasswordEncoder;
 import com.ddbb.dingdong.presentation.endpoint.auth.exchanges.SignUpRequestDto;
 import com.ddbb.dingdong.util.ParamValidator;
