+++ backend/src/main/java/com/ddbb/dingdong/infrastructure/auth/security/AuthUser.java
@@ -1,4 +1,4 @@
-package com.ddbb.dingdong.infrastructure.auth;
+package com.ddbb.dingdong.infrastructure.auth.security;
 
 import com.ddbb.dingdong.domain.user.entity.vo.Role;
