+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/admin/AdminController.java
@@ -6,8 +6,8 @@ import com.ddbb.dingdong.application.usecase.clustering.GetLocationUseCase;
 import com.ddbb.dingdong.application.usecase.routing.CreateRouteUseCase;
 import com.ddbb.dingdong.domain.common.exception.DomainException;
 import com.ddbb.dingdong.domain.reservation.entity.vo.Direction;
-import com.ddbb.dingdong.infrastructure.auth.AuthUser;
-import com.ddbb.dingdong.infrastructure.auth.annotation.LoginUser;
+import com.ddbb.dingdong.infrastructure.auth.security.AuthUser;
+import com.ddbb.dingdong.infrastructure.auth.security.annotation.LoginUser;
 import com.ddbb.dingdong.presentation.endpoint.admin.exchanges.CreateClusterRequestDTO;
 import com.ddbb.dingdong.presentation.endpoint.admin.exchanges.CreateRouteRequestDTO;
 import com.ddbb.dingdong.presentation.endpoint.admin.exchanges.CreateRouteResponseDTO;
