+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/admin/temp/AdminUserTestController.java
@@ -6,8 +6,8 @@ import com.ddbb.dingdong.application.usecase.user.FastSignUpUseCase;
 import com.ddbb.dingdong.domain.common.exception.DomainException;
 import com.ddbb.dingdong.domain.user.entity.User;
 import com.ddbb.dingdong.domain.user.repository.UserRepository;
-import com.ddbb.dingdong.infrastructure.auth.AuthUser;
-import com.ddbb.dingdong.infrastructure.auth.annotation.LoginUser;
+import com.ddbb.dingdong.infrastructure.auth.security.AuthUser;
+import com.ddbb.dingdong.infrastructure.auth.security.annotation.LoginUser;
 import com.ddbb.dingdong.presentation.endpoint.admin.temp.dto.FastSignUpDTO;
 import com.ddbb.dingdong.presentation.endpoint.admin.temp.dto.GetUserInfoDTO;
 import com.ddbb.dingdong.presentation.endpoint.admin.temp.dto.UserHomeUpdateDTO;
