+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/auth/AuthController.java
@@ -7,8 +7,8 @@ import com.ddbb.dingdong.application.usecase.auth.LogoutUseCase;
 import com.ddbb.dingdong.application.usecase.auth.SignUpUseCase;
 import com.ddbb.dingdong.domain.auth.service.error.AuthErrors;
 import com.ddbb.dingdong.domain.common.exception.DomainException;
-import com.ddbb.dingdong.infrastructure.auth.AuthUser;
-import com.ddbb.dingdong.infrastructure.auth.annotation.LoginUser;
+import com.ddbb.dingdong.infrastructure.auth.security.AuthUser;
+import com.ddbb.dingdong.infrastructure.auth.security.annotation.LoginUser;
 import com.ddbb.dingdong.presentation.endpoint.auth.exchanges.CheckEmailDto;
 import com.ddbb.dingdong.presentation.endpoint.auth.exchanges.SignUpRequestDto;
 import com.ddbb.dingdong.presentation.endpoint.auth.exchanges.LoginRequestDTO;
