+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/bus/BusController.java
@@ -7,8 +7,8 @@ import com.ddbb.dingdong.application.usecase.bus.GetBusStopLocationUseCase;
 import com.ddbb.dingdong.application.usecase.bus.GetPathPointLine;
 import com.ddbb.dingdong.domain.common.exception.DomainException;
 import com.ddbb.dingdong.domain.reservation.entity.vo.Direction;
-import com.ddbb.dingdong.infrastructure.auth.AuthUser;
-import com.ddbb.dingdong.infrastructure.auth.annotation.LoginUser;
+import com.ddbb.dingdong.infrastructure.auth.security.AuthUser;
+import com.ddbb.dingdong.infrastructure.auth.security.annotation.LoginUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
