+++ backend/src/main/java/com/ddbb/dingdong/presentation/endpoint/user/UserHomeController.java
@@ -4,8 +4,8 @@ import com.ddbb.dingdong.application.exception.APIException;
 import com.ddbb.dingdong.application.usecase.user.GetHomeLocationUseCase;
 import com.ddbb.dingdong.application.usecase.user.ChangeUserStationUseCase;
 import com.ddbb.dingdong.domain.common.exception.DomainException;
-import com.ddbb.dingdong.infrastructure.auth.AuthUser;
-import com.ddbb.dingdong.infrastructure.auth.annotation.LoginUser;
+import com.ddbb.dingdong.infrastructure.auth.security.AuthUser;
+import com.ddbb.dingdong.infrastructure.auth.security.annotation.LoginUser;
 import com.ddbb.dingdong.presentation.endpoint.user.exchanges.UpdateHomeLocationDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
